[feat_4.2.x_issue#116][common-loader-solr][支持slor]

+++ solr/src/main/java/com/dtstack/dtcenter/common/loader/solr/SolrUtils.java
@@ -1,10 +1,12 @@
 package com.dtstack.dtcenter.common.loader.solr;
 
+import com.dtstack.dtcenter.common.loader.hadoop.util.KerberosConfigUtil;
 import com.dtstack.dtcenter.loader.exception.DtLoaderException;
 import com.dtstack.dtcenter.loader.kerberos.HadoopConfTool;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections.MapUtils;
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.lang3.StringUtils;
 import sun.security.krb5.Config;
 
 import java.io.File;
@@ -46,18 +48,22 @@ public class SolrUtils {
             System.setProperty(HadoopConfTool.KEY_JAVA_SECURITY_KRB5_CONF, MapUtils.getString(kerberosConfig, HadoopConfTool.KEY_JAVA_SECURITY_KRB5_CONF));
         }
 
-        String keytabConf = MapUtils.getString(kerberosConfig, HadoopConfTool.PRINCIPAL_FILE);
+        String keytabPath = MapUtils.getString(kerberosConfig, HadoopConfTool.PRINCIPAL_FILE);
+        if (StringUtils.isBlank(keytabPath)) {
+            throw new DtLoaderException("Keytab path is null");
+        }
         try {
-            File file = new File(keytabConf);
+            File file = new File(keytabPath);
             File jaas = new File(file.getParent() + File.separator + "solr_jaas.conf");
             if (jaas.exists()) {
                 jaas.delete();
             }
 
-            String principal = MapUtils.getString(kerberosConfig, HadoopConfTool.PRINCIPAL);
-            FileUtils.write(jaas, String.format(SolrConsistent.SOLR_JAAS_CONTENT, keytabConf, principal));
+            //如果没有设置principal,默认读keytab取第一个
+            String principal = kerberosConfig.containsKey(HadoopConfTool.PRINCIPAL) ? MapUtils.getString(kerberosConfig, HadoopConfTool.PRINCIPAL) :  KerberosConfigUtil.getPrincipals(keytabPath).get(0);
+            FileUtils.write(jaas, String.format(SolrConsistent.SOLR_JAAS_CONTENT, keytabPath, principal));
             String solrLoginConf = jaas.getAbsolutePath();
-            log.info("Init solr Kerberos:login-conf:{}\n --sasl.kerberos.service.name:{}", keytabConf, principal);
+            log.info("Init solr Kerberos:login-conf:{}\n --principal:{}", keytabPath, principal);
             return solrLoginConf;
         } catch (IOException e) {
             throw new DtLoaderException(String.format("Writing to solr configuration file exception,%s", e.getMessage()), e);
